LayoutTests/editing/deleting/merge-into-empty-block-1-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of DIV > BODY > HTML > #document to 2 of DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldDeleteDOMRange:range from 0 of DIV > DIV > BODY > HTML > #document to 0 of BLOCKQUOTE > DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: webViewDidEndEditing:WebViewDidEndEditingNotification
When a user puts the caret at the very beginning of a blockquote and hits backspace, if the line before the blockquote is empty, the blockquote should just move up.
| <blockquote>
|   style="margin: 0px; padding-left: 25px; border-left: 3px solid blue; color: blue;"
|   type="cite"
|   <span>
|     id="test"
|     "<#selection-caret>foo"
